moved away from Dapper and static databases

+++ GameWatch.Core/Dto/ElapsedTime.cs
@@ -1,8 +1,8 @@
 ﻿namespace GameWatch.Core.Dto;
 
-public readonly record struct ElapsedTime(int V)
+public readonly record struct ElapsedTime(long V)
 {
-    public static readonly ElapsedTime Zero = new(0);
+    public static readonly ElapsedTime Zero = new(0L);
 
     public override string ToString() => V.ToString();
 }
